requirements.txt
# Core Dependencies

# New Packages (detected by pipreqs)
google_api_python_client==2.118.0
google_auth_oauthlib==1.2.0
markdown==3.7
orjson==3.10.15
pillow==11.1.0
protobuf==6.30.0
ujson==5.10.0
upstash_redis==1.3.0

numpy==1.24.3  # Must be installed first for spaCy compatibility

# Web Framework and ASGI/WSGI

Flask[async]==3.1.0
hypercorn==0.17.3
asgiref==3.8.1
Werkzeug>=3.1
Jinja2>=3.1.2
blinker>=1.9.0
click>=8.1.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0

# Email and Network

imapclient==3.0.1
aiohttp==3.11.11
httpx==0.28.1
httpcore==1.0.7
email-validator==2.1.0.post1
google-auth==2.27.0
google-auth-httplib2==0.2.0
requests-oauthlib>=1.3.1

# NLP and Machine Learning

spacy==3.6.1
en_core_web_sm@ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
openai==1.60.0
transformers==4.48.1
torch==2.5.1
sentence-transformers==3.3.1
tiktoken==0.6.0

# Data Processing and Validation

pydantic==2.10.5
python-dateutil==2.8.2
python-dotenv==1.0.1
tqdm==4.67.1
typing_extensions==4.12.2
dataclasses-json==0.6.4

# Development and Testing

pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock>=3.12.0
pytest-redis>=3.0.0
rich==13.9.4
coverage>=5.2.1

# Database and ORM

SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
Mako>=1.3.8
sqlalchemy-json>=0.7.0
sqlalchemy-utils>=0.41.0

# Caching and Performance

redis[hiredis]==5.0.1
hiredis>=2.3.0
async-timeout>=5.0.1

# Async Support

asyncio==3.4.3
aiosignal==1.3.2
aiohappyeyeballs==2.4.4
contextvars==2.4
propcache>=0.2.0
aiofiles>=23.2.1

# Logging and Monitoring

python-json-logger==2.0.7
structlog>=24.1.0

# Security and Authentication

itsdangerous>=2.2.0
pyOpenSSL==24.0.0
cryptography==42.0.8
oauthlib>=3.2.0
python-jose[cryptography]>=3.3.0

# Utilities

six>=1.5
packaging>=20.0
setuptools>=75.8.0
pyparsing!=3.0.0,!=3.0.1,!=3.0.2,!=3.0.3,<4,>=2.4.2
MarkupSafe>=2.0
cloudpathlib<1.0.0,>=0.7.0
smart-open<8.0.0,>=5.2.1
wrapt>=1.17.2
requests==2.31.0
httplib2>=0.19.0
python-slugify>=8.0.0
python-magic>=0.4.27
psutil==5.9.8

# Documentation tools

Sphinx>=8.2.0
furo>=2024.8.0
interrogate>=1.5.0
myst-parser>=4.0.0

# Script utilities (previously in scripts/requirements.txt)
astroid>=2.8.0
ast-comments>=1.0.1
pathlib>=1.0.1  # Note: May be redundant as pathlib is part of Python's standard library in Python 3.4+
